split utils into utils and ldap, pull request changes
