+++ Profiles/__init__.py
@@ -5,7 +5,6 @@ import csh_ldap
 
 import flask_migrate
 from flask import Flask, render_template, jsonify, request, redirect, send_from_directory
-from flask_optimize import FlaskOptimize
 from flask_pyoidc.flask_pyoidc import OIDCAuthentication
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import func
@@ -13,8 +12,6 @@ from sqlalchemy import func
 
 app = Flask(__name__)
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-
-flask_optimize = FlaskOptimize()
 
 # Get app config from absolute file path
 if os.path.exists(os.path.join(os.getcwd(), "config.py")):
@@ -38,28 +35,18 @@ requests.packages.urllib3.disable_warnings()
 # LDAP
 ldap = csh_ldap.CSHLDAP(app.config['LDAP_BIND_DN'], app.config['LDAP_BIND_PASS'])
 
-from Profiles.utils import before_request
-from Profiles.utils import ldap_get_active_members
-from Profiles.utils import ldap_get_all_members
-from Profiles.utils import ldap_get_member
-from Profiles.utils import ldap_search_members
-from Profiles.utils import ldap_is_active
-from Profiles.utils import ldap_get_eboard
-from Profiles.utils import get_member_info
-from Profiles.utils import _ldap_get_group_members
+from Profiles.utils import before_request, get_member_info
+from Profiles.ldap import ldap_get_active_members, ldap_get_all_members, ldap_get_member, ldap_search_members, ldap_is_active, ldap_get_eboard, _ldap_get_group_members
 
 @app.route("/", methods=["GET"])
 @auth.oidc_auth
-@flask_optimize.optimize()
 @before_request
 def home(info=None):
-    return render_template("profile.html", 
-    						  info=info, 
-    						  member_info=info["member_info"])
+    return redirect("/profile/" + info["uid"],
+                              code = 302)
 
 @app.route("/members", methods=["GET"])
 @auth.oidc_auth
-@flask_optimize.optimize()
 @before_request
 def members(info=None):
     return render_template("members.html", 
@@ -69,7 +56,6 @@ def members(info=None):
 
 @app.route("/profile/<uid>", methods=["GET"])
 @auth.oidc_auth
-@flask_optimize.optimize()
 @before_request
 def profile(uid=None, info=None):
     return render_template("profile.html", 
@@ -78,7 +64,6 @@ def profile(uid=None, info=None):
 
 @app.route("/results", methods=["POST"])
 @auth.oidc_auth
-@flask_optimize.optimize()
 @before_request
 def results(uid=None, info=None):
     if request.method == "POST":
@@ -90,7 +75,6 @@ def results(uid=None, info=None):
 
 @app.route("/search/<searched>", methods=["GET"])
 @auth.oidc_auth
-@flask_optimize.optimize()
 @before_request
 def search(searched=None, info=None):
     return render_template("members.html", 
@@ -100,12 +84,11 @@ def search(searched=None, info=None):
 
 @app.route("/group/<group>", methods=["GET"])
 @auth.oidc_auth
-@flask_optimize.optimize()
 @before_request
 def group(group=None, info=None):
     if "eboard" in group:
     	return render_template("members.html", 
-    						    info=info, 
+    						    info=info,
     						    title = "Group: " + group,
     						    members=ldap_get_eboard())
     else:
